Add LocalStorage testing
