+++ package.json
@@ -21,6 +21,11 @@
 		"url": "https://github.com/carlosmuerto/to-do-list-webpack/issues"
 	},
 	"homepage": "https://github.com/carlosmuerto/to-do-list-webpack#readme",
+	"jest": {
+		"globals": {
+			"window": {}
+		}
+	},
 	"devDependencies": {
 		"@babel/plugin-transform-modules-commonjs": "^7.18.6",
 		"babel-eslint": "^10.1.0",
